test: 💍 add test for REAL chain test
